Finished the runner script
# All the bugs seem to be fixed.

Maven really doesn’t like for non local files to be used. Thus I now copy the file in if it’s already extracted or find the correct file in a tar.gz and extract it in. Next the analysis is run. Then the output is exported to the correct directory and then I clean up the tempory files.

Guess that will do the trick?